requirements.txt
flask==0.12
Redis
cassandra-driver
Jinja2==2.9.5
flask_bootstrap